third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-auto-columns-rows-get-set-expected.txt

Test that setting and getting grid-auto-columns and grid-auto-rows works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting grid-auto-columns and grid-auto-rows set through CSS
PASS getComputedStyle(gridAutoFixedFixed, '').getPropertyValue('grid-auto-rows') is '30px'
PASS getComputedStyle(gridAutoFixedFixed, '').getPropertyValue('grid-auto-columns') is '50px'
PASS getComputedStyle(gridAutoMinMax, '').getPropertyValue('grid-auto-rows') is 'minmax(10%, 15px)'
PASS getComputedStyle(gridAutoMinMax, '').getPropertyValue('grid-auto-columns') is 'minmax(30%, 100px)'
PASS getComputedStyle(gridAutoMinMaxContent, '').getPropertyValue('grid-auto-rows') is 'min-content'
PASS getComputedStyle(gridAutoMinMaxContent, '').getPropertyValue('grid-auto-columns') is 'max-content'

Test the initial value
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is 'auto'

Test getting and setting grid-auto-columns and grid-auto-rows through JS
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is '180px'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is '660px'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is 'minmax(min-content, 48px)'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is 'minmax(80px, min-content)'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is 'minmax(min-content, max-content)'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is 'minmax(max-content, min-content)'

Test setting grid-auto-columns and grid-auto-rows to bad minmax value through JS
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is 'auto'

Test setting grid-auto-columns and grid-auto-rows to 'inherit' through JS
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is '50px'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is '101%'

Test setting grid-auto-columns and grid-auto-rows to 'initial' through JS
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is '150%'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is '1fr'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is 'auto'
PASS successfullyParsed is true

TEST COMPLETE
